+++ src/controls/select.ts
@@ -39,6 +39,7 @@ declare global {
       showCase: ShowCases;
     }
     interface Options<T = any, VM = ValidityMap> extends WUP.BaseCombo.Options<T, VM>, Defaults<T, VM> {
+      // todo need somehow notify Proxy not to wrap the Array: same behavior for radio
       /** Items showed in dropdown-menu. Provide promise/api-call to show pending status when control retrieves data! */
       items: MenuItems<T> | (() => MenuItems<T> | Promise<MenuItems<T>>) | Promise<MenuItems<T>>;
       /** Set true to make input not editable but allow to user select items via popup-menu (ordinary dropdown mode)
@@ -790,3 +791,4 @@ customElements.define(tagName, WUPSelectControl);
 
 // NiceToHave: option to allow autoselect item without pressing Enter
 // WARN Chrome touchscreen simulation issue: touch on label>strong fires click on input - the issue only in simulation
+// todo search value by {id} for complex object - the same for radio